+++ packages/react/src/components/PdfViewer/__tests__/PdfViewer.basic.test.tsx
@@ -1,6 +1,6 @@
 import { cleanup, render, screen, waitFor } from "@testing-library/react";
 import { afterEach, expect, test } from "vitest";
-import { PdfViewer } from "./index";
+import { PdfViewer } from "../index";
 
 afterEach(cleanup);
